for review only
please do not use this repo  for anything, it was forked from original
naut and i manually overwrote naut 2 onto it to review all the changes.
Yes I realize there are other more efficient ways to review separate
repositories.  When a project starts on one repository it should stay
there. Different trees can be used for upgrades or different versions
but creating a new repository is imo an improper way to show your
changes to the world from one version to another.